# Faster package querying