--- a/compose/cli/__init__.py
+++ b/compose/cli/__init__.py
@@ -17,12 +17,15 @@ try:
     env[str('PIP_DISABLE_PIP_VERSION_CHECK')] = str('1')
 
     s_cmd = subprocess.Popen(
-        ['pip', 'freeze'], stderr=subprocess.PIPE, stdout=subprocess.PIPE,
+        [
+            sys.executable, '-c',
+            'import pkg_resources; print("\\n".join(map(str, [distr for distr in pkg_resources.working_set if distr.project_name])))'
+        ], stderr=subprocess.PIPE, stdout=subprocess.PIPE,
         env=env
     )
     packages = s_cmd.communicate()[0].splitlines()
     dockerpy_installed = len(
-        list(filter(lambda p: p.startswith(b'docker-py=='), packages))
+        list(filter(lambda p: p.startswith(b'docker-py '), packages))
     ) > 0
     if dockerpy_installed:
         from .colors import yellow