During the `docker-py` distribution check, `pip freeze` can be terribly slow if you have a lot of packages installed. Here's my case:

    $ time pip freeze > /dev/null
    real    0m3,930s
    user    0m3,787s
     sys     0m0,093s

    $ pip freeze | wc -l
    173

    $ time docker-compose 2> /dev/null
    real    0m3,443s
    user    0m3,277s
    sys     0m0,143s

After this fix it becomes more than 3x faster:

    $ time docker-compose 2> /dev/null 
    real    0m0,935s
    user    0m0,827s
    sys     0m0,100s

This feature was tested in Python 2.7 & 3.6.