@and3rson 2017-07-06
The Jenkins build seems to be hanging... Is it normal?