@shin- 2017-07-07
https://github.com/docker/compose/pull/4920#issuecomment-310774092