@shin- 2017-07-07
This part of the code has been iterated over many times. The current solution is the one that offers the best compromise for every declination of the software.